fix(twine): change batch and block numbers to bigint

+++ apps/explorer/priv/twine/migrations/20250221093546_create_twine_tables.exs
@@ -3,9 +3,9 @@ defmodule Explorer.Repo.Twine.Migrations.CreateTwineTables do
 
   def change do
     create table(:twine_transaction_batch, primary_key: false) do
-      add(:number, :integer, null: false, primary_key: true)
-      add(:start_block, :integer, null: false)
-      add(:end_block, :integer, null: false)
+      add(:number, :bigint, null: false, primary_key: true)
+      add(:start_block, :bigint, null: false)
+      add(:end_block, :bigint, null: false)
       add(:timestamp, :utc_datetime_usec, null: false)
       add(:root_hash, :bytea, null: false)
       timestamps(null: false, type: :utc_datetime_usec)
@@ -13,7 +13,7 @@ defmodule Explorer.Repo.Twine.Migrations.CreateTwineTables do
 
     create table(:twine_transaction_batch_detail, primary_key: false) do
       add(:id, :serial, null: false, primary_key: true)
-      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :delete_all, on_update: :update_all, type: :integer), null: false)
+      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :delete_all, on_update: :update_all, type: :bigint), null: false)
       add(:l1_transaction_count, :integer, null: false)
       add(:l2_transaction_count, :integer, null: false)
       add(:l1_gas_price, :numeric, precision: 100, null: false)
@@ -42,14 +42,14 @@ defmodule Explorer.Repo.Twine.Migrations.CreateTwineTables do
 
     # Add batch_number to transactions table
     alter table(:transactions) do
-      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :nilify_all, on_update: :update_all, type: :integer), null: true)
+      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :nilify_all, on_update: :update_all, type: :bigint), null: true)
     end
 
     create(index(:transactions, :batch_number))
 
     # Add batch_number to blocks table
     alter table(:blocks) do
-      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :nilify_all, on_update: :update_all, type: :integer), null: true)
+      add(:batch_number, references(:twine_transaction_batch, column: :number, on_delete: :nilify_all, on_update: :update_all, type: :bigint), null: true)
     end
 
     create(index(:blocks, :batch_number))
